refactor player screen into modular components
